refactor(src): moved BackgroundMap into component dir
